Remove the separate dialog for choosing the briefcase storage location.
